+++ api/configuration/appConfig.ts
@@ -1,44 +1,58 @@
-import * as express from 'express'
-import {getConfigValue, showFeature} from '../configuration'
+import {getConfigValue, showFeature} from './index'
+
 import {
+  COOKIES_SESSION_ID,
+  COOKIES_TOKEN,
   FEATURE_APP_INSIGHTS_ENABLED,
   FEATURE_PROXY_ENABLED,
   FEATURE_REDIS_ENABLED,
   FEATURE_SECURE_COOKIE_ENABLED,
   FEATURE_TERMS_AND_CONDITIONS_ENABLED,
+  HEALTH,
+  LOGGING,
+  MAX_LOG_LINE,
   PROTOCOL,
   SERVICE_S2S_PATH,
   SERVICES_CCD_COMPONENT_API_PATH,
   SERVICES_CCD_DATA_STORE_API_PATH,
   SERVICES_DOCUMENTS_API_PATH,
+  SERVICES_IDAM_API_URL,
   SERVICES_IDAM_CLIENT_ID,
   SERVICES_IDAM_LOGIN_URL,
   SERVICES_IDAM_OAUTH_CALLBACK_URL,
-} from '../configuration/references'
-
-export const router = express.Router({mergeParams: true})
-
-router.get('/', configurationCheckRoute)
+  SERVICES_TERMS_AND_CONDITIONS_URL,
+} from './references'
 
 /**
- * Used to check application configuration
+ * Application Configuration
+ *
+ * Details the applications configuration.
+ *
+ * Note that we do not include the secrets within this object.
+ *
+ * @returns
  */
-async function configurationCheckRoute(req, res) {
-  res.status(200).send({
+export const applicationConfiguration = () => {
+  return {
     clientId: getConfigValue(SERVICES_IDAM_CLIENT_ID),
+    cookieSessionId: getConfigValue(COOKIES_SESSION_ID),
+    cookieToken: getConfigValue(COOKIES_TOKEN),
     featureAppInsightsEnabled: showFeature(FEATURE_APP_INSIGHTS_ENABLED),
     featureProxyEnabled: showFeature(FEATURE_PROXY_ENABLED),
     featureRedisEnabled: showFeature(FEATURE_REDIS_ENABLED),
     featureSecureCookieEnabled: showFeature(FEATURE_SECURE_COOKIE_ENABLED),
     featureTermsAndConditionsEnabled: showFeature(FEATURE_TERMS_AND_CONDITIONS_ENABLED),
+    health: getConfigValue(HEALTH),
+    idamApiUrl: getConfigValue(SERVICES_IDAM_API_URL),
     idamWeb: getConfigValue(SERVICES_IDAM_LOGIN_URL),
+    logging: getConfigValue(LOGGING),
+    maxLogLine: getConfigValue(MAX_LOG_LINE),
     oAuthCallback: getConfigValue(SERVICES_IDAM_OAUTH_CALLBACK_URL),
     protocol: getConfigValue(PROTOCOL),
     servicesCcdComponentPath: getConfigValue(SERVICES_CCD_COMPONENT_API_PATH),
     servicesCcdDataStorePath: getConfigValue(SERVICES_CCD_DATA_STORE_API_PATH),
     servicesDocumentApiPath: getConfigValue(SERVICES_DOCUMENTS_API_PATH),
     servicesS2sPath: getConfigValue(SERVICE_S2S_PATH),
-  })
+    servicesTermsAndConditionsPath: getConfigValue(SERVICES_TERMS_AND_CONDITIONS_URL),
+  }
 }
-
-export default router
