+++ api/configuration/uiConfigRouter.ts
@@ -1,20 +1,25 @@
 import * as express from 'express'
-import {getConfigValue} from '../configuration'
+import {getConfigValue} from './index'
 import {
   PROTOCOL,
   SERVICES_IDAM_CLIENT_ID,
   SERVICES_IDAM_LOGIN_URL,
   SERVICES_IDAM_OAUTH_CALLBACK_URL,
-} from '../configuration/references'
+} from './references'
 
 export const router = express.Router({mergeParams: true})
 
-router.get('/', configurationUIRoute)
+router.get('/', uiConfigurationRouter)
 
 /**
- * All the following environmental variables are passed to the UI.
+ * UI Configuration Route
+ *
+ * The following environmental variables are passed to the UI.
+ *
+ * Note that we do not pass the whole appConfiguration to the UI,
+ * we only pass what's needed to be precise.
  */
-async function configurationUIRoute(req, res) {
+async function uiConfigurationRouter(req, res) {
   res.status(200).send({
     clientId: getConfigValue(SERVICES_IDAM_CLIENT_ID),
     idamWeb: getConfigValue(SERVICES_IDAM_LOGIN_URL),
