+++ config/custom-environment-variables.json
@@ -31,7 +31,6 @@
       "idamApiUrl": "SERVICES_IDAM_API_URL",
       "idamClientID": "SERVICES_IDAM_CLIENT_ID",
       "idamLoginUrl": "SERVICES_IDAM_LOGIN_URL",
-      "indexUrl": "SERVICES_IDAM_INDEX_URL",
       "oauthCallbackUrl": "SERVICES_IDAM_OAUTH_CALLBACK_URL"
     },
     "s2s": "SERVICES_S2S",
